refactor: move WSEditor.tsx into own dir
