+++ src/web/src/views/workspace/WSEditor/index.ts
@@ -1,0 +1,5 @@
+export { WSEditor, WSEditor as default } from "./WSEditor";
+export { default as WSEditorExportDialog } from "./WSEditorExportDialog";
+export { default as WSEditorDeleteDialog } from "./WSEditorDeleteDialog";
+export { default as WSEditorSwaggerReloadDialog } from "./WSEditorSwaggerReloadDialog";
+export { default as WSRenameDialog } from "./WSRenameDialog";
